move autoparameter methods to separate file
